# generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fnirs-flow API",
    "description": "fNIRS analysis Flow orchestration framework",
    "version": "1.3.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/bundle": {
      "get": {
        "summary": "Project Bundle Status Endpoint",
        "operationId": "project_bundle_status_endpoint_api_projects__project_id__bundle_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/bundle/restore/{revision}": {
      "post": {
        "summary": "Restore Project Bundle Endpoint",
        "operationId": "restore_project_bundle_endpoint_api_projects__project_id__bundle_restore__revision__post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "revision",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/version-history": {
      "get": {
        "summary": "Get Version History Endpoint",
        "operationId": "get_version_history_endpoint_api_projects__project_id__version_history_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VersionHistoryEntry"
                  },
                  "title": "Response Get Version History Endpoint Api Projects  Project Id  Version History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/export-package": {
      "post": {
        "summary": "Export Package Endpoint",
        "operationId": "export_package_endpoint_api_projects__project_id__export_package_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ExportRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{project_id}/import-package": {
      "post": {
        "summary": "Import Package Endpoint",
        "operationId": "import_package_endpoint_api_projects__project_id__import_package_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "package_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Path"
            }
          },
          {
            "name": "data_root",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Root"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Import Package Endpoint Api Projects  Project Id  Import Package Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/fork": {
      "post": {
        "summary": "Fork Project Endpoint",
        "operationId": "fork_project_endpoint_api_projects__project_id__fork_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "fork_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Fork Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Fork Project Endpoint Api Projects  Project Id  Fork Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/trust-atom/{atom_id}": {
      "post": {
        "summary": "Trust Atom Endpoint",
        "operationId": "trust_atom_endpoint_api_projects__project_id__trust_atom__atom_id__post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "atom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Atom Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Trust Atom Endpoint Api Projects  Project Id  Trust Atom  Atom Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/import-status": {
      "get": {
        "summary": "Import Status Endpoint",
        "operationId": "import_status_endpoint_api_projects__project_id__import_status_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/relink-data": {
      "post": {
        "summary": "Relink Project Data Endpoint",
        "operationId": "relink_project_data_endpoint_api_projects__project_id__relink_data_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "data_root",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Data Root"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Relink Project Data Endpoint Api Projects  Project Id  Relink Data Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/bind-dataset": {
      "post": {
        "summary": "Bind Project Dataset Endpoint",
        "operationId": "bind_project_dataset_endpoint_api_projects__project_id__bind_dataset_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "data_root",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Data Root"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Bind Project Dataset Endpoint Api Projects  Project Id  Bind Dataset Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BundleStatus": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "package_path": {
            "type": "string",
            "title": "Package Path",
            "default": ""
          },
          "storage_format": {
            "type": "string",
            "title": "Storage Format",
            "default": "fnirsflow_bundle"
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 0
          },
          "saved_at": {
            "type": "string",
            "title": "Saved At",
            "default": ""
          },
          "save_reason": {
            "type": "string",
            "title": "Save Reason",
            "default": ""
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes",
            "default": 0
          },
          "integrity_status": {
            "type": "string",
            "title": "Integrity Status",
            "default": "unknown"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          },
          "verification_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Scope"
          },
          "integrity_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integrity Error"
          },
          "dirty": {
            "type": "boolean",
            "title": "Dirty",
            "default": false
          },
          "busy_operation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Busy Operation"
          },
          "lock_owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lock Owner"
          },
          "versions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "BundleStatus",
        "description": "Extended bundle status with integrity and lock information."
      },
...
      "ExportRequest": {
        "properties": {
          "profile": {
            "type": "string",
            "title": "Profile",
            "default": "reproducibility_package"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "attempt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id"
          },
          "include_history": {
            "type": "boolean",
            "title": "Include History",
            "default": false
          }
        },
        "type": "object",
        "title": "ExportRequest"
      },
      "ExportResult": {
        "properties": {
          "package_path": {
            "type": "string",
            "title": "Package Path"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "profile": {
            "type": "string",
            "title": "Profile",
            "default": "reproducibility_package"
          },
          "contents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contents"
          }
        },
        "type": "object",
        "required": [
          "package_path",
          "size_bytes"
        ],
        "title": "ExportResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportStatus": {
        "properties": {
          "imported": {
            "type": "boolean",
            "title": "Imported"
          },
          "read_only": {
            "type": "boolean",
            "title": "Read Only"
          },
          "quarantined_atoms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Quarantined Atoms"
          },
          "relinked": {
            "type": "boolean",
            "title": "Relinked",
            "default": false
          },
          "data_root": {
            "type": "string",
            "title": "Data Root",
            "default": ""
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "imported",
          "read_only"
        ],
        "title": "ImportStatus"
      },
...
      "ProjectRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "data_root": {
            "type": "string",
            "title": "Data Root",
            "default": ""
          },
          "flow_id": {
            "type": "string",
            "title": "Flow Id",
            "default": ""
          },
          "package_path": {
            "type": "string",
            "title": "Package Path",
            "default": ""
          },
          "storage_format": {
            "type": "string",
            "title": "Storage Format",
            "default": "fnirsflow_bundle"
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 0
          },
          "integrity_status": {
            "type": "string",
            "title": "Integrity Status",
            "default": "unknown"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          },
          "verification_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Scope"
          },
          "integrity_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integrity Error"
          },
          "busy_operation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Busy Operation"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "ProjectRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VersionHistoryEntry": {
        "properties": {
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "saved_at": {
            "type": "string",
            "title": "Saved At"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "current": {
            "type": "boolean",
            "title": "Current"
          },
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "revision",
          "saved_at",
          "reason",
          "current",
          "path"
        ],
        "title": "VersionHistoryEntry"
      }
    }
  }
}